fix: refactor settings
